修改数据库表，替换旧的数据表 修改数据表字段 student_new 测试通过

+++ src/main/java/yangchen/exam/entity/ExamInfo.java
@@ -19,7 +19,7 @@ public class ExamInfo {
     private Integer id;
 
     @Column(name = "student_number")
-    private Long studentNumber;
+    private Integer studentNumber;
 
     //学生姓名
     @Column(name = "student_name")
